+++ fme/ace/aggregator/inference/enso/__init__.py
@@ -1,6 +1,2 @@
-from .dynamic_index import (
-    LatLonRegion,
-    PairedRegionalIndexAggregator,
-    RegionalIndexAggregator,
-)
+from .dynamic_index import PairedRegionalIndexAggregator, RegionalIndexAggregator
 from .enso_coefficient import EnsoCoefficientEvaluatorAggregator
